// src/app/common/enum/filter-type.ts
export enum FilterType {
  stringContains = 'stringContains',
  numberEqual = 'numberEqual',
  numberMoreThan = 'numberMoreThan',
  numberLessThan = 'numberLessThan',
  numberMoreOrEqualThan = 'numberMoreOrEqualThan',
  numberLessOrEqualThan = 'numberLessOrEqualThan',
  numberContaine = 'numberContaine',
}